refactor: restructure codebase to new folder architecture
