+++ examples/arbitrage-bot.ts
@@ -1,0 +1,6 @@
+/**
+ * Example: Arbitrage bot
+ */
+
+// TODO: Implement example
+console.log('Arbitrage bot example');
