+++ examples/simple-binance-bot.ts
@@ -1,0 +1,6 @@
+/**
+ * Example: Simple Binance bot
+ */
+
+// TODO: Implement example
+console.log('Simple Binance bot example');
