+++ src/connectors/base-connector.ts
@@ -1,0 +1,8 @@
+/**
+ * Base connector interface for WaspBot-TS
+ */
+
+// TODO: Implement base connector
+export abstract class BaseConnector {
+  // Placeholder
+}
