+++ src/connectors/binance.ts
@@ -1,0 +1,8 @@
+/**
+ * Binance connector for WaspBot-TS
+ */
+
+// TODO: Implement Binance connector
+export class BinanceConnector {
+  // Placeholder
+}
