+++ src/connectors/kucoin.ts
@@ -1,0 +1,8 @@
+/**
+ * KuCoin connector for WaspBot-TS
+ */
+
+// TODO: Implement KuCoin connector
+export class KuCoinConnector {
+  // Placeholder
+}
