+++ src/connectors/uniswap.ts
@@ -1,0 +1,8 @@
+/**
+ * Uniswap connector for WaspBot-TS
+ */
+
+// TODO: Implement Uniswap connector
+export class UniswapConnector {
+  // Placeholder
+}
