+++ src/core/clock.ts
@@ -1,0 +1,9 @@
+/**
+ * @fileoverview Clock and scheduling system for WaspBot-TS
+ * Handles event/tick scheduling
+ */
+
+// TODO: Implement clock system
+export class Clock {
+  // Placeholder
+}
