+++ src/core/logger.ts
@@ -1,0 +1,8 @@
+/**
+ * @fileoverview Logging utility for WaspBot-TS
+ */
+
+// TODO: Implement logging
+export class Logger {
+  // Placeholder
+}
