+++ src/core/state-manager.ts
@@ -1,0 +1,8 @@
+/**
+ * @fileoverview Shared bot state manager for WaspBot-TS
+ */
+
+// TODO: Implement state management
+export class StateManager {
+  // Placeholder
+}
