+++ src/strategies/arbitrage.ts
@@ -1,0 +1,8 @@
+/**
+ * Arbitrage strategy for WaspBot-TS
+ */
+
+// TODO: Implement arbitrage strategy
+export class ArbitrageStrategy {
+  // Placeholder
+}
