+++ src/strategies/base-strategy.ts
@@ -1,0 +1,8 @@
+/**
+ * Base strategy interface for WaspBot-TS
+ */
+
+// TODO: Implement base strategy
+export abstract class BaseStrategy {
+  // Placeholder
+}
