+++ src/strategies/ping-pong.ts
@@ -1,0 +1,8 @@
+/**
+ * Ping-pong strategy for WaspBot-TS
+ */
+
+// TODO: Implement ping-pong strategy
+export class PingPongStrategy {
+  // Placeholder
+}
