+++ src/strategies/trend-following.ts
@@ -1,0 +1,8 @@
+/**
+ * Trend-following strategy for WaspBot-TS
+ */
+
+// TODO: Implement trend-following strategy
+export class TrendFollowingStrategy {
+  // Placeholder
+}
