+++ src/utils/http-client.ts
@@ -1,0 +1,8 @@
+/**
+ * HTTP client utility for WaspBot-TS
+ */
+
+// TODO: Implement HTTP client
+export class HttpClient {
+  // Placeholder
+}
