+++ src/utils/math.ts
@@ -1,0 +1,8 @@
+/**
+ * Math utilities for WaspBot-TS
+ */
+
+// TODO: Implement math utils
+export class MathUtils {
+  // Placeholder
+}
