+++ src/utils/ws-client.ts
@@ -1,0 +1,8 @@
+/**
+ * WebSocket client utility for WaspBot-TS
+ */
+
+// TODO: Implement WS client
+export class WsClient {
+  // Placeholder
+}
