+++ tests/connectors.test.ts
@@ -1,0 +1,10 @@
+/**
+ * Unit tests for connectors
+ */
+
+// TODO: Implement connector tests
+describe('Connectors', () => {
+  it('should work', () => {
+    // Placeholder
+  });
+});
