+++ tests/order-manager.test.ts
@@ -1,0 +1,10 @@
+/**
+ * Unit tests for order manager
+ */
+
+// TODO: Implement order manager tests
+describe('OrderManager', () => {
+  it('should work', () => {
+    // Placeholder
+  });
+});
