+++ tests/strategies.test.ts
@@ -1,0 +1,10 @@
+/**
+ * Unit tests for strategies
+ */
+
+// TODO: Implement strategy tests
+describe('Strategies', () => {
+  it('should work', () => {
+    // Placeholder
+  });
+});
